Remove all row vectors, removed templating on derived base everywhere.

+++ mav_trajectory_generation/src/segment.cpp
@@ -34,7 +34,7 @@ const Polynomial& Segment::operator[](size_t idx) const {
 
 Eigen::VectorXd Segment::evaluate(double t, int derivative) const {
   Eigen::VectorXd result(D_);
-  for (size_t d = 0; d < D_; ++d) {
+  for (int d = 0; d < D_; ++d) {
     result[d] = polynomials_[d].evaluate(t, derivative);
   }
   return result;
@@ -45,7 +45,7 @@ void printSegment(std::ostream& stream, const Segment& s, int derivative) {
   stream << "t: " << s.getTime() << std::endl;
   stream << " coefficients for " << positionDerivativeToString(derivative)
          << ": " << std::endl;
-  for (size_t i = 0; i < s.D(); ++i) {
+  for (int i = 0; i < s.D(); ++i) {
     stream << s[i].getCoefficients(derivative) << std::endl;
   }
 }
